[WIP] build: allow statically linking geos
Release note: None
