+++ c-deps/buildcdeps.sh
@@ -16,7 +16,7 @@ for CONFIG in $CONFIGS; do
         TARGETS="$TARGETS libkrb5"
     fi
     bazel clean
-    bazel build --config ci --config cross$CONFIG --//build/toolchains:prebuild_cdeps_flag $(echo "$TARGETS" | python3 -c 'import sys; input = sys.stdin.read().strip(); print(" ".join("//c-deps:{}_foreign".format(w) for w in input.split(" ")))')
+    bazel build --config ci --config cross$CONFIG --//build/toolchains:prebuild_cdeps_flag --//build/toolchains:geos_dynamic_flag=false $(echo "$TARGETS" | python3 -c 'import sys; input = sys.stdin.read().strip(); print(" ".join("//c-deps:{}_foreign".format(w) for w in input.split(" ")))')
     BAZEL_BIN=$(bazel info bazel-bin --config ci --config cross$CONFIG)
     for TARGET in $TARGETS; do
         # verify jemalloc was configured without madv_free
@@ -29,6 +29,9 @@ for CONFIG in $CONFIGS; do
         fi
         bundle $CONFIG $BAZEL_BIN/c-deps/${TARGET}_foreign
     done
+    # Also take care of shared/dynamic library geos.
+    bazel build --config ci --config cross$CONFIG --//build/toolchains:prebuild_cdeps_flag --//build/toolchains:geos_dynamic_flag=true //c-deps:libgeos_foreign
+    bundle ${CONFIG}_libgeos_shared $BAZEL_BIN/c-deps/libgeos_foreign
 done
 
 for FILE in $(find /artifacts -name '*.tar.gz' | sort); do
