+++ pkg/geo/geos/geos.cc
@@ -10,11 +10,6 @@
 
 #include <cstdarg>
 #include <cstring>
-#if _WIN32
-#include <windows.h>
-#else
-#include <dlfcn.h>
-#endif  // #if _WIN32
 #include <memory>
 #include <stdio.h>
 #include <stdlib.h>
@@ -22,387 +17,13 @@
 #include <vector>
 
 #include "geos.h"
-
-#if _WIN32
-#define dlopen(x, y) LoadLibrary(x)
-#define dlsym GetProcAddress
-#define dlclose FreeLibrary
-#define dlerror() ((char*)"failed to execute dlsym")
-typedef HMODULE dlhandle;
-#else
-typedef void* dlhandle;
-#endif  // #if _WIN32
+#include "geos_cc.hh"
 
 #define CR_GEOS_NO_ERROR_DEFINED_MESSAGE "geos: returned invalid result but error not populated"
 
 namespace {
-
-// Data Types adapted from `capi/geos_c.h.in` in GEOS.
-typedef void* CR_GEOS_Handle;
-typedef void (*CR_GEOS_MessageHandler)(const char*, void*);
-typedef void* CR_GEOS_WKTReader;
-typedef void* CR_GEOS_WKBReader;
-typedef void* CR_GEOS_WKBWriter;
-typedef void* CR_GEOS_BufferParams;
-
-// Function declarations from `capi/geos_c.h.in` in GEOS.
-typedef CR_GEOS_Handle (*CR_GEOS_init_r)();
-typedef void (*CR_GEOS_finish_r)(CR_GEOS_Handle);
-typedef CR_GEOS_MessageHandler (*CR_GEOS_Context_setErrorMessageHandler_r)(CR_GEOS_Handle,
-                                                                           CR_GEOS_MessageHandler,
-                                                                           void*);
-typedef void (*CR_GEOS_Free_r)(CR_GEOS_Handle, void* buffer);
-typedef void (*CR_GEOS_SetSRID_r)(CR_GEOS_Handle, CR_GEOS_Geometry, int);
-typedef int (*CR_GEOS_GetSRID_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-typedef void (*CR_GEOS_GeomDestroy_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-
-typedef char (*CR_GEOS_HasZ_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-typedef char (*CR_GEOS_IsEmpty_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-typedef char (*CR_GEOS_IsSimple_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-typedef int (*CR_GEOS_GeomTypeId_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-
-typedef CR_GEOS_WKTReader (*CR_GEOS_WKTReader_create_r)(CR_GEOS_Handle);
-typedef CR_GEOS_Geometry (*CR_GEOS_WKTReader_read_r)(CR_GEOS_Handle, CR_GEOS_WKTReader,
-                                                     const char*);
-typedef void (*CR_GEOS_WKTReader_destroy_r)(CR_GEOS_Handle, CR_GEOS_WKTReader);
-
-typedef CR_GEOS_WKBReader (*CR_GEOS_WKBReader_create_r)(CR_GEOS_Handle);
-typedef CR_GEOS_Geometry (*CR_GEOS_WKBReader_read_r)(CR_GEOS_Handle, CR_GEOS_WKBReader, const char*,
-                                                     size_t);
-typedef void (*CR_GEOS_WKBReader_destroy_r)(CR_GEOS_Handle, CR_GEOS_WKBReader);
-
-typedef CR_GEOS_BufferParams (*CR_GEOS_BufferParams_create_r)(CR_GEOS_Handle);
-typedef void (*CR_GEOS_GEOSBufferParams_destroy_r)(CR_GEOS_Handle, CR_GEOS_BufferParams);
-typedef int (*CR_GEOS_BufferParams_setEndCapStyle_r)(CR_GEOS_Handle, CR_GEOS_BufferParams,
-                                                     int endCapStyle);
-typedef int (*CR_GEOS_BufferParams_setJoinStyle_r)(CR_GEOS_Handle, CR_GEOS_BufferParams,
-                                                   int joinStyle);
-typedef int (*CR_GEOS_BufferParams_setMitreLimit_r)(CR_GEOS_Handle, CR_GEOS_BufferParams,
-                                                    double mitreLimit);
-typedef int (*CR_GEOS_BufferParams_setQuadrantSegments_r)(CR_GEOS_Handle, CR_GEOS_BufferParams,
-                                                          int quadrantSegments);
-typedef int (*CR_GEOS_BufferParams_setSingleSided_r)(CR_GEOS_Handle, CR_GEOS_BufferParams,
-                                                     int singleSided);
-typedef CR_GEOS_Geometry (*CR_GEOS_BufferWithParams_r)(CR_GEOS_Handle, CR_GEOS_Geometry,
-                                                       CR_GEOS_BufferParams, double width);
-
-typedef char (*CR_GEOS_isValid_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-typedef char* (*CR_GEOS_isValidReason_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-typedef char (*CR_GEOS_isValidDetail_r)(CR_GEOS_Handle, CR_GEOS_Geometry, int flags, char** reason,
-                                        CR_GEOS_Geometry* loc);
-typedef CR_GEOS_Geometry (*CR_GEOS_MakeValid_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-
-typedef int (*CR_GEOS_Area_r)(CR_GEOS_Handle, CR_GEOS_Geometry, double*);
-typedef int (*CR_GEOS_Length_r)(CR_GEOS_Handle, CR_GEOS_Geometry, double*);
-typedef int (*CR_GEOS_MinimumClearance_r)(CR_GEOS_Handle, CR_GEOS_Geometry, double*);
-typedef CR_GEOS_Geometry (*CR_GEOS_MinimumClearanceLine_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-typedef int (*CR_GEOS_Normalize_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-typedef CR_GEOS_Geometry (*CR_GEOS_LineMerge_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-
-typedef CR_GEOS_Geometry (*CR_GEOS_Boundary_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-typedef CR_GEOS_Geometry (*CR_GEOS_Centroid_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-typedef CR_GEOS_Geometry (*CR_GEOS_ConvexHull_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-typedef CR_GEOS_Geometry (*CR_GEOS_Difference_r)(CR_GEOS_Handle, CR_GEOS_Geometry,
-                                                 CR_GEOS_Geometry);
-typedef CR_GEOS_Geometry (*CR_GEOS_Simplify_r)(CR_GEOS_Handle, CR_GEOS_Geometry, double);
-typedef CR_GEOS_Geometry (*CR_GEOS_TopologyPreserveSimplify_r)(CR_GEOS_Handle, CR_GEOS_Geometry,
-                                                               double);
-typedef CR_GEOS_Geometry (*CR_GEOS_UnaryUnion_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-typedef CR_GEOS_Geometry (*CR_GEOS_Union_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry);
-typedef CR_GEOS_Geometry (*CR_GEOS_Intersection_r)(CR_GEOS_Handle, CR_GEOS_Geometry,
-                                                   CR_GEOS_Geometry);
-typedef CR_GEOS_Geometry (*CR_GEOS_SymDifference_r)(CR_GEOS_Handle, CR_GEOS_Geometry,
-                                                    CR_GEOS_Geometry);
-typedef CR_GEOS_Geometry (*CR_GEOS_PointOnSurface_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-
-typedef CR_GEOS_Geometry (*CR_GEOS_Interpolate_r)(CR_GEOS_Handle, CR_GEOS_Geometry, double);
-
-typedef CR_GEOS_Geometry (*CR_GEOS_MinimumBoundingCircle_r)(CR_GEOS_Handle, CR_GEOS_Geometry, double*, CR_GEOS_Geometry**);
-
-typedef int (*CR_GEOS_Distance_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry, double*);
-typedef int (*CR_GEOS_FrechetDistance_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry,
-                                         double*);
-typedef int (*CR_GEOS_FrechetDistanceDensify_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry,
-                                                double, double*);
-typedef int (*CR_GEOS_HausdorffDistance_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry,
-                                           double*);
-typedef int (*CR_GEOS_HausdorffDistanceDensify_r)(CR_GEOS_Handle, CR_GEOS_Geometry,
-                                                  CR_GEOS_Geometry, double, double*);
-
-typedef CR_GEOS_PreparedInternalGeometry (*CR_GEOS_Prepare_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-typedef void (*CR_GEOS_PreparedGeom_destroy_r)(CR_GEOS_Handle, CR_GEOS_PreparedInternalGeometry);
-
-typedef char (*CR_GEOS_PreparedIntersects_r)(CR_GEOS_Handle, CR_GEOS_PreparedInternalGeometry, CR_GEOS_Geometry);
-
-typedef char (*CR_GEOS_Covers_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry);
-typedef char (*CR_GEOS_CoveredBy_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry);
-typedef char (*CR_GEOS_Contains_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry);
-typedef char (*CR_GEOS_Crosses_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry);
-typedef char (*CR_GEOS_Disjoint_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry);
-typedef char (*CR_GEOS_Equals_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry);
-typedef char (*CR_GEOS_Intersects_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry);
-typedef char (*CR_GEOS_Overlaps_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry);
-typedef char (*CR_GEOS_Touches_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry);
-typedef char (*CR_GEOS_Within_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry);
-
-typedef char* (*CR_GEOS_Relate_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry);
-typedef char* (*CR_GEOS_RelateBoundaryNodeRule_r)(CR_GEOS_Handle, CR_GEOS_Geometry,
-                                                  CR_GEOS_Geometry, int);
-typedef char (*CR_GEOS_RelatePattern_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry,
-                                        const char*);
-
-typedef CR_GEOS_WKBWriter (*CR_GEOS_WKBWriter_create_r)(CR_GEOS_Handle);
-typedef char* (*CR_GEOS_WKBWriter_write_r)(CR_GEOS_Handle, CR_GEOS_WKBWriter, CR_GEOS_Geometry,
-                                           size_t*);
-typedef void (*CR_GEOS_WKBWriter_setByteOrder_r)(CR_GEOS_Handle, CR_GEOS_WKBWriter, int);
-typedef void (*CR_GEOS_WKBWriter_setOutputDimension_r)(CR_GEOS_Handle, CR_GEOS_WKBWriter, int);
-typedef void (*CR_GEOS_WKBWriter_destroy_r)(CR_GEOS_Handle, CR_GEOS_WKBWriter);
-typedef void (*CR_GEOS_WKBWriter_setIncludeSRID_r)(CR_GEOS_Handle, CR_GEOS_WKBWriter, const char);
-typedef CR_GEOS_Geometry (*CR_GEOS_ClipByRect_r)(CR_GEOS_Handle, CR_GEOS_Geometry, double, double,
-                                                 double, double);
-
-typedef CR_GEOS_Geometry (*CR_GEOS_SharedPaths_r)(CR_GEOS_Handle, CR_GEOS_Geometry,
-                                                  CR_GEOS_Geometry);
-
-typedef CR_GEOS_Geometry (*CR_GEOS_Node_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-
-typedef CR_GEOS_Geometry (*CR_GEOS_VoronoiDiagram_r)(CR_GEOS_Handle, CR_GEOS_Geometry,
-                                                  CR_GEOS_Geometry, double, int);
-
-typedef char (*CR_GEOS_EqualsExact_r)(CR_GEOS_Handle, CR_GEOS_Geometry,
-                                                  CR_GEOS_Geometry, double);
-
-typedef CR_GEOS_Geometry (*CR_GEOS_MinimumRotatedRectangle_r)(CR_GEOS_Handle, CR_GEOS_Geometry);
-
-typedef CR_GEOS_Geometry (*CR_GEOS_Snap_r)(CR_GEOS_Handle, CR_GEOS_Geometry, CR_GEOS_Geometry, double);
-
-std::string ToString(CR_GEOS_Slice slice) { return std::string(slice.data, slice.len); }
-
-}  // namespace
-
-struct CR_GEOS {
-  dlhandle geoscHandle;
-  dlhandle geosHandle;
-
-  CR_GEOS_init_r GEOS_init_r;
-  CR_GEOS_finish_r GEOS_finish_r;
-  CR_GEOS_Context_setErrorMessageHandler_r GEOSContext_setErrorMessageHandler_r;
-  CR_GEOS_Free_r GEOSFree_r;
-
-  CR_GEOS_HasZ_r GEOSHasZ_r;
-  CR_GEOS_IsEmpty_r GEOSisEmpty_r;
-  CR_GEOS_IsSimple_r GEOSisSimple_r;
-  CR_GEOS_GeomTypeId_r GEOSGeomTypeId_r;
-
-  CR_GEOS_BufferParams_create_r GEOSBufferParams_create_r;
-  CR_GEOS_GEOSBufferParams_destroy_r GEOSBufferParams_destroy_r;
-  CR_GEOS_BufferParams_setEndCapStyle_r GEOSBufferParams_setEndCapStyle_r;
-  CR_GEOS_BufferParams_setJoinStyle_r GEOSBufferParams_setJoinStyle_r;
-  CR_GEOS_BufferParams_setMitreLimit_r GEOSBufferParams_setMitreLimit_r;
-  CR_GEOS_BufferParams_setQuadrantSegments_r GEOSBufferParams_setQuadrantSegments_r;
-  CR_GEOS_BufferParams_setSingleSided_r GEOSBufferParams_setSingleSided_r;
-  CR_GEOS_BufferWithParams_r GEOSBufferWithParams_r;
-
-  CR_GEOS_SetSRID_r GEOSSetSRID_r;
-  CR_GEOS_GetSRID_r GEOSGetSRID_r;
-  CR_GEOS_GeomDestroy_r GEOSGeom_destroy_r;
-
-  CR_GEOS_WKTReader_create_r GEOSWKTReader_create_r;
-  CR_GEOS_WKTReader_destroy_r GEOSWKTReader_destroy_r;
-  CR_GEOS_WKTReader_read_r GEOSWKTReader_read_r;
-
-  CR_GEOS_WKBReader_create_r GEOSWKBReader_create_r;
-  CR_GEOS_WKBReader_destroy_r GEOSWKBReader_destroy_r;
-  CR_GEOS_WKBReader_read_r GEOSWKBReader_read_r;
-
-  CR_GEOS_isValid_r GEOSisValid_r;
-  CR_GEOS_isValidReason_r GEOSisValidReason_r;
-  CR_GEOS_isValidDetail_r GEOSisValidDetail_r;
-  CR_GEOS_MakeValid_r GEOSMakeValid_r;
-
-  CR_GEOS_Area_r GEOSArea_r;
-  CR_GEOS_Length_r GEOSLength_r;
-  CR_GEOS_MinimumClearance_r GEOSMinimumClearance_r;
-  CR_GEOS_MinimumClearanceLine_r GEOSMinimumClearanceLine_r;
-  CR_GEOS_Normalize_r GEOSNormalize_r;
-  CR_GEOS_LineMerge_r GEOSLineMerge_r;
-
-  CR_GEOS_Boundary_r GEOSBoundary_r;
-  CR_GEOS_Centroid_r GEOSGetCentroid_r;
-  CR_GEOS_ConvexHull_r GEOSConvexHull_r;
-  CR_GEOS_Difference_r GEOSDifference_r;
-  CR_GEOS_Simplify_r GEOSSimplify_r;
-  CR_GEOS_TopologyPreserveSimplify_r GEOSTopologyPreserveSimplify_r;
-  CR_GEOS_UnaryUnion_r GEOSUnaryUnion_r;
-  CR_GEOS_Union_r GEOSUnion_r;
-  CR_GEOS_PointOnSurface_r GEOSPointOnSurface_r;
-  CR_GEOS_Intersection_r GEOSIntersection_r;
-  CR_GEOS_SymDifference_r GEOSSymDifference_r;
-
-  CR_GEOS_MinimumBoundingCircle_r GEOSMinimumBoundingCircle_r;
-
-  CR_GEOS_Interpolate_r GEOSInterpolate_r;
-
-  CR_GEOS_Distance_r GEOSDistance_r;
-  CR_GEOS_FrechetDistance_r GEOSFrechetDistance_r;
-  CR_GEOS_FrechetDistanceDensify_r GEOSFrechetDistanceDensify_r;
-  CR_GEOS_HausdorffDistance_r GEOSHausdorffDistance_r;
-  CR_GEOS_HausdorffDistanceDensify_r GEOSHausdorffDistanceDensify_r;
-
-  CR_GEOS_Prepare_r GEOSPrepare_r;
-  CR_GEOS_PreparedGeom_destroy_r GEOSPreparedGeom_destroy_r;
-
-  CR_GEOS_PreparedIntersects_r GEOSPreparedIntersects_r;
-
-  CR_GEOS_Covers_r GEOSCovers_r;
-  CR_GEOS_CoveredBy_r GEOSCoveredBy_r;
-  CR_GEOS_Contains_r GEOSContains_r;
-  CR_GEOS_Crosses_r GEOSCrosses_r;
-  CR_GEOS_Disjoint_r GEOSDisjoint_r;
-  CR_GEOS_Equals_r GEOSEquals_r;
-  CR_GEOS_Intersects_r GEOSIntersects_r;
-  CR_GEOS_Overlaps_r GEOSOverlaps_r;
-  CR_GEOS_Touches_r GEOSTouches_r;
-  CR_GEOS_Within_r GEOSWithin_r;
-
-  CR_GEOS_Relate_r GEOSRelate_r;
-  CR_GEOS_RelateBoundaryNodeRule_r GEOSRelateBoundaryNodeRule_r;
-  CR_GEOS_RelatePattern_r GEOSRelatePattern_r;
-
-  CR_GEOS_WKBWriter_create_r GEOSWKBWriter_create_r;
-  CR_GEOS_WKBWriter_destroy_r GEOSWKBWriter_destroy_r;
-  CR_GEOS_WKBWriter_setByteOrder_r GEOSWKBWriter_setByteOrder_r;
-  CR_GEOS_WKBWriter_setOutputDimension_r GEOSWKBWriter_setOutputDimension_r;
-  CR_GEOS_WKBWriter_setIncludeSRID_r GEOSWKBWriter_setIncludeSRID_r;
-  CR_GEOS_WKBWriter_write_r GEOSWKBWriter_write_r;
-
-  CR_GEOS_ClipByRect_r GEOSClipByRect_r;
-
-  CR_GEOS_SharedPaths_r GEOSSharedPaths_r;
-  CR_GEOS_VoronoiDiagram_r GEOSVoronoiDiagram_r;
-  CR_GEOS_EqualsExact_r GEOSEqualsExact_r;
-  CR_GEOS_MinimumRotatedRectangle_r GEOSMinimumRotatedRectangle_r;
-
-  CR_GEOS_Node_r GEOSNode_r;
-
-  CR_GEOS_Snap_r GEOSSnap_r;
-
-  CR_GEOS(dlhandle geoscHandle, dlhandle geosHandle)
-      : geoscHandle(geoscHandle), geosHandle(geosHandle) {}
-
-  ~CR_GEOS() {
-    if (geoscHandle != NULL) {
-      dlclose(geoscHandle);
-    }
-    if (geosHandle != NULL) {
-      dlclose(geosHandle);
-    }
-  }
-
-  char* Init() {
-#define INIT(x)                                                                                    \
-  do {                                                                                             \
-    auto error = InitSym(&x, #x);                                                                  \
-    if (error != nullptr) {                                                                        \
-      return error;                                                                                \
-    }                                                                                              \
-  } while (0)
-
-    INIT(GEOS_init_r);
-    INIT(GEOS_finish_r);
-    INIT(GEOSFree_r);
-    INIT(GEOSContext_setErrorMessageHandler_r);
-    INIT(GEOSGeom_destroy_r);
-    INIT(GEOSBufferParams_create_r);
-    INIT(GEOSBufferParams_destroy_r);
-    INIT(GEOSBufferParams_setEndCapStyle_r);
-    INIT(GEOSBufferParams_setJoinStyle_r);
-    INIT(GEOSBufferParams_setMitreLimit_r);
-    INIT(GEOSBufferParams_setQuadrantSegments_r);
-    INIT(GEOSBufferParams_setSingleSided_r);
-    INIT(GEOSBufferWithParams_r);
-    INIT(GEOSHasZ_r);
-    INIT(GEOSisEmpty_r);
-    INIT(GEOSGeomTypeId_r);
-    INIT(GEOSSetSRID_r);
-    INIT(GEOSGetSRID_r);
-    INIT(GEOSisValid_r);
-    INIT(GEOSisValidReason_r);
-    INIT(GEOSisValidDetail_r);
-    INIT(GEOSMakeValid_r);
-    INIT(GEOSArea_r);
-    INIT(GEOSLength_r);
-    INIT(GEOSMinimumClearance_r);
-    INIT(GEOSMinimumClearanceLine_r);
-    INIT(GEOSNormalize_r);
-    INIT(GEOSLineMerge_r);
-    INIT(GEOSisSimple_r);
-    INIT(GEOSBoundary_r);
-    INIT(GEOSDifference_r);
-    INIT(GEOSGetCentroid_r);
-    INIT(GEOSMinimumBoundingCircle_r);
-    INIT(GEOSConvexHull_r);
-    INIT(GEOSSimplify_r);
-    INIT(GEOSTopologyPreserveSimplify_r);
-    INIT(GEOSUnaryUnion_r);
-    INIT(GEOSUnion_r);
-    INIT(GEOSPointOnSurface_r);
-    INIT(GEOSIntersection_r);
-    INIT(GEOSSymDifference_r);
-    INIT(GEOSInterpolate_r);
-    INIT(GEOSDistance_r);
-    INIT(GEOSFrechetDistance_r);
-    INIT(GEOSFrechetDistanceDensify_r);
-    INIT(GEOSHausdorffDistance_r);
-    INIT(GEOSHausdorffDistanceDensify_r);
-    INIT(GEOSPrepare_r);
-    INIT(GEOSPreparedGeom_destroy_r);
-    INIT(GEOSPreparedIntersects_r);
-    INIT(GEOSCovers_r);
-    INIT(GEOSCoveredBy_r);
-    INIT(GEOSContains_r);
-    INIT(GEOSCrosses_r);
-    INIT(GEOSDisjoint_r);
-    INIT(GEOSEquals_r);
-    INIT(GEOSIntersects_r);
-    INIT(GEOSOverlaps_r);
-    INIT(GEOSTouches_r);
-    INIT(GEOSWithin_r);
-    INIT(GEOSRelate_r);
-    INIT(GEOSVoronoiDiagram_r);
-    INIT(GEOSEqualsExact_r);
-    INIT(GEOSMinimumRotatedRectangle_r);
-    INIT(GEOSRelateBoundaryNodeRule_r);
-    INIT(GEOSRelatePattern_r);
-    INIT(GEOSSharedPaths_r);
-    INIT(GEOSWKTReader_create_r);
-    INIT(GEOSWKTReader_destroy_r);
-    INIT(GEOSWKTReader_read_r);
-    INIT(GEOSWKBReader_create_r);
-    INIT(GEOSWKBReader_destroy_r);
-    INIT(GEOSWKBReader_read_r);
-    INIT(GEOSWKBWriter_create_r);
-    INIT(GEOSWKBWriter_destroy_r);
-    INIT(GEOSWKBWriter_setByteOrder_r);
-    INIT(GEOSWKBWriter_setOutputDimension_r);
-    INIT(GEOSWKBWriter_setIncludeSRID_r);
-    INIT(GEOSWKBWriter_write_r);
-    INIT(GEOSClipByRect_r);
-    INIT(GEOSNode_r);
-    INIT(GEOSSnap_r);
-    return nullptr;
-
-#undef INIT
-  }
-
-  template <typename T> char* InitSym(T* ptr, const char* symbol) {
-    *ptr = reinterpret_cast<T>(dlsym(geoscHandle, symbol));
-    if (ptr == nullptr) {
-      return dlerror();
-    }
-    return nullptr;
-  }
-};
+  std::string ToString(CR_GEOS_Slice slice) { return std::string(slice.data, slice.len); }
+}
 
 CR_GEOS_Slice cStringToSlice(char* cstr) {
   CR_GEOS_Slice slice = {.data = cstr, .len = strlen(cstr)};
@@ -424,37 +45,6 @@ CR_GEOS_String toGEOSString(const char* data, size_t len) {
 void errorHandler(const char* msg, void* buffer) {
   std::string* str = static_cast<std::string*>(buffer);
   *str = std::string("geos error: ") + msg;
-}
-
-CR_GEOS_Status CR_GEOS_Init(CR_GEOS_Slice geoscLoc, CR_GEOS_Slice geosLoc, CR_GEOS** lib) {
-  // Open the libgeos.$(EXT) first, so that libgeos_c.$(EXT) can read it.
-  std::string error;
-  auto geosLocStr = ToString(geosLoc);
-  dlhandle geosHandle = dlopen(geosLocStr.c_str(), RTLD_LAZY);
-  if (!geosHandle) {
-    errorHandler(dlerror(), &error);
-    return toGEOSString(error.data(), error.length());
-  }
-
-  auto geoscLocStr = ToString(geoscLoc);
-  dlhandle geoscHandle = dlopen(geoscLocStr.c_str(), RTLD_LAZY);
-  if (!geoscHandle) {
-    errorHandler(dlerror(), &error);
-    dlclose(geosHandle);
-    return toGEOSString(error.data(), error.length());
-  }
-
-  std::unique_ptr<CR_GEOS> ret(new CR_GEOS(geoscHandle, geosHandle));
-  auto initError = ret->Init();
-  if (initError != nullptr) {
-    errorHandler(initError, &error);
-    dlclose(geosHandle);
-    dlclose(geoscHandle);
-    return toGEOSString(error.data(), error.length());
-  }
-
-  *lib = ret.release();
-  return toGEOSString(error.data(), error.length());
 }
 
 CR_GEOS_Handle initHandleWithErrorBuffer(CR_GEOS* lib, std::string* buffer) {
